More cleanup, ponder what to do about RagRatQuery
I don't want to step on toes, this is a big change.
It kinda needs to be big tho..
There's still some little problems like I don't fully understand memories so it's only retrieved and not saved?
I want to try to keep functions that were around, or at least offer full replacement